# application/api-spec/src/main/resources/openapi/food-rescue.yaml
openapi: 3.0.3
info:
  title: Savr Platform API
  version: 0.0.1-SNAPSHOT
  description: |
    REST API for the Savr platform.
servers:
  - url: http://localhost:8080
...
paths:
...
  /businesses:
    get:
      tags:
        - Business
      summary: List businesses of current user
      description: |
        Returns all businesses owned by the authenticated user.

      operationId: getBusinesses
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Businesses listed successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BusinessResponse'
              examples:
                businessesFound:
                  value:
                    - id: 10
                      ownerId: 2
                      name: Bistro DobrÃ© Jedlo
                      description: Vegetarian rescue meals
                      status: PENDING
                      address:
                        street: Hlavna 1
                        city: Kosice
                        postalCode: 04001
                        country: Slovakia
                      createdAt: 2025-04-08T18:30:00Z
                    - id: 11
                      ownerId: 2
                      name: Green Kitchen
                      description: Fresh rescue lunches
                      status: ACTIVE
                      address:
                        street: Mestska 5
                        city: Kosice
                        postalCode: 04001
                        country: Slovakia
                      createdAt: 2025-04-10T10:00:00Z
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'

    post:
      tags:
        - Business
      summary: Create a business
      description: |
        Creates a new business.

      operationId: createBusiness
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: Business creation payload
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBusinessRequest'
            examples:
              businessCreate:
                summary: New business
                value:
                  name: Bistro Dobré Jedlo
                  description: Vegetarian rescue meals
                  address:
                    street: Hlavna 1
                    city: Kosice
                    postalCode: 04001
                    country: Slovakia
      responses:
        '201':
          description: Business created successfully
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: User is blocked or not allowed to create a business
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                blockedUser:
                  value:
                    code: FORBIDDEN
                    message: Access denied
                    details:
                      - Insufficient permissions
                    timestamp: 2025-04-08T18:30:00Z
                    path: /businesses
        '409':
          description: A business with the same name already exists for the owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicateBusiness:
                  value:
                    code: CONFLICT
                    message: Business with the same name already exists for this owner
                    details: []
                    timestamp: 2025-04-08T18:30:00Z
                    path: /businesses
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /businesses/{id}:
    get:
      tags:
        - Business
      summary: Get business by id
      description: |
        Returns a business by id.

      operationId: getBusiness
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BusinessIdPathParam'
      responses:
        '200':
          description: Business found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessResponse'
              examples:
                businessFound:
                  value:
                    id: 10
                    ownerId: 2
                    name: Bistro Dobré Jedlo
                    description: Vegetarian rescue meals
                    status: PENDING
                    address:
                      street: Hlavna 1
                      city: Kosice
                      postalCode: 04001
                      country: Slovakia
                    createdAt: 2025-04-08T18:30:00Z
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: User is not allowed to access this business
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                forbiddenBusiness:
                  value:
                    code: FORBIDDEN
                    message: You do not have access to business with id=10
                    details: []
                    timestamp: 2025-04-08T18:30:00Z
                    path: /businesses/10
        '404':
          description: Business not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                businessNotFound:
                  value:
                    code: NOT_FOUND
                    message: Business with id=10 was not found
                    details: []
                    timestamp: 2025-04-08T18:30:00Z
                    path: /businesses/10
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'

    put:
      tags:
        - Business
      summary: Update business
      description: |
        Updates a business owned by the authenticated user.

      operationId: updateBusiness
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BusinessIdPathParam'
      requestBody:
        required: true
        description: Business update payload
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBusinessRequest'
            examples:
              businessUpdate:
                summary: Updated business
                value:
                  name: Updated Bistro Dobré Jedlo
                  description: Updated vegetarian rescue meals
                  address:
                    street: Hlavna 12
                    city: Kosice
                    postalCode: 04001
                    country: Slovakia
      responses:
        '200':
          description: Business updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: User is not allowed to update this business
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                forbiddenUpdate:
                  value:
                    code: FORBIDDEN
                    message: You do not have access to business with id=10
                    details: []
                    timestamp: 2025-04-08T18:30:00Z
                    path: /businesses/10
        '404':
          description: Business not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: A business with the same name already exists for the owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicateBusinessOnUpdate:
                  value:
                    code: CONFLICT
                    message: Business with the same name already exists for this owner
                    details: []
                    timestamp: 2025-04-08T18:30:00Z
                    path: /businesses/10
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'

    delete:
      tags:
        - Business
      summary: Delete business
      description: |
        Deletes a business.

      operationId: deleteBusiness
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BusinessIdPathParam'
      responses:
        '204':
          description: Business deleted successfully
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: User is not allowed to delete this business
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                forbiddenDelete:
                  value:
                    code: FORBIDDEN
                    message: You do not have access to business with id=10
                    details: []
                    timestamp: 2025-04-08T18:30:00Z
                    path: /businesses/10
        '404':
          description: Business not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /businesses/{id}/analytics:
    get:
      tags:
        - Business
      summary: Get business analytics
      description: |
        Returns analytics for a business managed by the authenticated user.

      operationId: getBusinessAnalytics
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BusinessIdPathParam'
      responses:
        '200':
          description: Business analytics returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessAnalyticsResponse'
              examples:
                businessAnalytics:
                  value:
                    businessId: 10
                    businessName: Bistro Dobre Jedlo
                    generatedAt: 2025-04-28T16:30:00Z
                    overview:
                      totalOffers: 18
                      availableOffers: 3
                      offersClaimed: 14
                      activeReservations: 2
                      completedPickups: 10
                      cancelledReservations: 2
                      recoveredRevenue: 49.00
                      pendingReservedRevenue: 9.80
                      offerClaimRate: 77.8
                      pickupSuccessRate: 71.4
                      cancellationRate: 14.3
                      averageHoursToFirstReservation: 4.5
                    catalogStatus:
                      - status: AVAILABLE
                        count: 3
                        share: 16.7
                      - status: RESERVED
                        count: 5
                        share: 27.8
                    dailyActivity:
                      - date: 2025-04-27
                        offersPublished: 2
                        reservationsCreated: 3
                        cancellations: 0
                        pickupsConfirmed: 1
                    daypartPerformance:
                      - slotKey: evening
                        label: Evening
                        offersScheduled: 8
                        offersClaimed: 7
                        pickupsConfirmed: 5
                        recoveredRevenue: 24.50
                        claimRate: 87.5
                        pickupRate: 62.5
                    topItems:
                      - itemName: Croissant
                        offeredQuantity: 22
                        reservedQuantity: 18
                        pickedUpQuantity: 12
                    insights:
                      - tone: success
                        title: Reserved revenue is already in motion
                        message: 9.80 EUR are currently tied to active reservations waiting for pickup.
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: User is not allowed to access analytics for this business
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                forbiddenAnalytics:
                  value:
                    code: FORBIDDEN
                    message: You are not allowed to view analytics for this business
                    details: []
                    timestamp: 2025-04-08T18:30:00Z
                    path: /businesses/10/analytics
        '404':
          description: Business not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                analyticsBusinessNotFound:
                  value:
                    code: NOT_FOUND
                    message: Business with id=10 was not found
                    details: []
                    timestamp: 2025-04-08T18:30:00Z
                    path: /businesses/10/analytics
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /admin/businesses/pending:
    get:
      tags:
        - Admin
      summary: List pending businesses for approval
      description: |
        Returns businesses waiting for administrative approval.

      operationId: getPendingBusinessesForApproval
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pending businesses listed successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BusinessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: Admin privileges required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /admin/businesses/{id}/approve:
    post:
      tags:
        - Admin
      summary: Approve a business
      description: |
        Approves a pending business and moves it into the active state.

      operationId: approveBusiness
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BusinessIdPathParam'
      responses:
        '200':
          description: Business approved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: Admin privileges required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Business not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BusinessIdPathParam:
      name: id
      in: path
      required: true
      description: Business identifier
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 10
...
  responses:
    BadRequestResponse:
      description: Invalid request body or request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidBody:
              value:
                code: VALIDATION_ERROR
                message: Invalid request body
                details:
                  - name must not be blank
                timestamp: 2025-04-08T18:30:00Z
                path: /businesses
    UnauthorizedResponse:
      description: Authentication required or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                code: UNAUTHORIZED
                message: Full authentication is required to access this resource
                details: []
                timestamp: 2025-04-08T18:30:00Z
                path: /businesses
    InternalServerErrorResponse:
      description: Unexpected internal error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internalError:
              value:
                code: INTERNAL_ERROR
                message: Unexpected internal error
                details:
                  - Please contact support
                timestamp: 2025-04-08T18:30:00Z
                path: /businesses
  schemas:
...
    CreateBusinessRequest:
      type: object
      required:
        - name
        - address
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: Bistro Dobré Jedlo
        description:
          type: string
          nullable: true
          maxLength: 1000
          example: Vegetarian rescue meals
        address:
          $ref: '#/components/schemas/Address'
    UpdateBusinessRequest:
      type: object
      required:
        - name
        - address
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: Updated Bistro Dobré Jedlo
        description:
          type: string
          nullable: true
          maxLength: 1000
          example: Updated vegetarian rescue meals
        address:
          $ref: '#/components/schemas/Address'
    BusinessResponse:
      type: object
      required:
        - id
        - ownerId
        - name
        - status
        - address
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          example: 10
        ownerId:
          type: integer
          format: int64
          example: 2
        name:
          type: string
          example: Bistro Dobré Jedlo
        description:
          type: string
          nullable: true
          example: Vegetarian rescue meals
        status:
          $ref: '#/components/schemas/BusinessStatus'
        address:
          $ref: '#/components/schemas/Address'
        createdAt:
          type: string
          format: date-time
          example: 2025-04-08T18:30:00Z
    BusinessAnalyticsResponse:
      type: object
      required:
        - businessId
        - businessName
        - generatedAt
        - overview
        - catalogStatus
        - dailyActivity
        - daypartPerformance
        - topItems
        - insights
      properties:
        businessId:
          type: integer
          format: int64
          example: 10
        businessName:
          type: string
          example: Bistro Dobre Jedlo
        generatedAt:
          type: string
          format: date-time
          example: 2025-04-28T16:30:00Z
        overview:
          $ref: '#/components/schemas/BusinessAnalyticsOverview'
        catalogStatus:
          type: array
          items:
            $ref: '#/components/schemas/BusinessAnalyticsCatalogStatus'
        dailyActivity:
          type: array
          items:
            $ref: '#/components/schemas/BusinessAnalyticsDailyActivityPoint'
        daypartPerformance:
          type: array
          items:
            $ref: '#/components/schemas/BusinessAnalyticsDaypartPerformance'
        topItems:
          type: array
          items:
            $ref: '#/components/schemas/BusinessAnalyticsItemPerformance'
        insights:
          type: array
          items:
            $ref: '#/components/schemas/BusinessAnalyticsInsight'
    BusinessAnalyticsOverview:
      type: object
      required:
        - totalOffers
        - availableOffers
        - offersClaimed
        - activeReservations
        - completedPickups
        - cancelledReservations
        - recoveredRevenue
        - pendingReservedRevenue
        - offerClaimRate
        - pickupSuccessRate
        - cancellationRate
      properties:
        totalOffers:
          type: integer
          example: 18
        availableOffers:
          type: integer
          example: 3
        offersClaimed:
          type: integer
          example: 14
        activeReservations:
          type: integer
          example: 2
        completedPickups:
          type: integer
          example: 10
        cancelledReservations:
          type: integer
          example: 2
        recoveredRevenue:
          type: number
          format: double
          example: 49.00
        pendingReservedRevenue:
          type: number
          format: double
          example: 9.80
        offerClaimRate:
          type: number
          format: double
          example: 77.8
        pickupSuccessRate:
          type: number
          format: double
          example: 71.4
        cancellationRate:
          type: number
          format: double
          example: 14.3
        averageHoursToFirstReservation:
          type: number
          format: double
          nullable: true
          example: 4.5
    BusinessAnalyticsCatalogStatus:
      type: object
      required:
        - status
        - count
        - share
      properties:
        status:
          type: string
          example: AVAILABLE
        count:
          type: integer
          example: 3
        share:
          type: number
          format: double
          example: 16.7
    BusinessAnalyticsDailyActivityPoint:
      type: object
      required:
        - date
        - offersPublished
        - reservationsCreated
        - cancellations
        - pickupsConfirmed
      properties:
        date:
          type: string
          format: date
          example: 2025-04-27
        offersPublished:
          type: integer
          example: 2
        reservationsCreated:
          type: integer
          example: 3
        cancellations:
          type: integer
          example: 0
        pickupsConfirmed:
          type: integer
          example: 1
    BusinessAnalyticsDaypartPerformance:
      type: object
      required:
        - slotKey
        - label
        - offersScheduled
        - offersClaimed
        - pickupsConfirmed
        - recoveredRevenue
        - claimRate
        - pickupRate
      properties:
        slotKey:
          type: string
          example: evening
        label:
          type: string
          example: Evening
        offersScheduled:
          type: integer
          example: 8
        offersClaimed:
          type: integer
          example: 7
        pickupsConfirmed:
          type: integer
          example: 5
        recoveredRevenue:
          type: number
          format: double
          example: 24.50
        claimRate:
          type: number
          format: double
          example: 87.5
        pickupRate:
          type: number
          format: double
          example: 62.5
    BusinessAnalyticsItemPerformance:
      type: object
      required:
        - itemName
        - offeredQuantity
        - reservedQuantity
        - pickedUpQuantity
      properties:
        itemName:
          type: string
          example: Croissant
        offeredQuantity:
          type: integer
          example: 22
        reservedQuantity:
          type: integer
          example: 18
        pickedUpQuantity:
          type: integer
          example: 12
    BusinessAnalyticsInsight:
      type: object
      required:
        - tone
        - title
        - message
      properties:
        tone:
          type: string
          example: success
        title:
          type: string
          example: Reserved revenue is already in motion
        message:
          type: string
          example: 9.80 EUR are currently tied to active reservations waiting for pickup.
...
    Address:
      type: object
      required:
        - street
        - city
        - postalCode
        - country
      properties:
        street:
          type: string
          minLength: 1
          maxLength: 255
          example: Hlavna 1
        city:
          type: string
          minLength: 1
          maxLength: 100
          example: Kosice
        postalCode:
          type: string
          minLength: 1
          maxLength: 20
          example: 04001
        country:
          type: string
          minLength: 1
          maxLength: 100
          example: Slovakia
        latitude:
          type: number
          format: double
          nullable: true
          minimum: -90
          maximum: 90
          example: 48.7205
        longitude:
          type: number
          format: double
          nullable: true
          minimum: -180
          maximum: 180
          example: 21.2578
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - details
        - timestamp
        - path
      properties:
        code:
          type: string
          example: CONFLICT
        message:
          type: string
          example: Business with the same name already exists for this owner
        details:
          type: array
          items:
            type: string
          example:
            - Duplicate or invalid database state
        timestamp:
          type: string
          format: date-time
          example: 2025-04-08T18:30:00Z
        path:
          type: string
          example: /businesses
...
    BusinessStatus:
      type: string
      enum:
        - PENDING
        - ACTIVE
        - BLOCKED
        - REJECTED
...